callback added

+++ Hippo/Wrappers/FuguConfig/HippoConfig.swift
@@ -156,7 +156,7 @@ struct BotAction {
     public var HippoLanguageChanged : ((Error?)->())?
     public var HippoSessionStatus: ((GroupCallStatus)->())?
     public var announcementUnreadCount : ((Int)->())?
-    public var hideTabbar : (()->())?
+    public var hideTabbar : ((Int)->())?
     
     internal let powererdByColor = #colorLiteral(red: 0.4980392157, green: 0.4980392157, blue: 0.4980392157, alpha: 1)
     internal let FuguColor = #colorLiteral(red: 0.3843137255, green: 0.4901960784, blue: 0.8823529412, alpha: 1)
